Move SecurityHeaderMiddleware to Services namespace
